refactor(frontend): page-title composition moves into @agenta/shared
The title rules — normalize, truncate on a character boundary, join the parts —
lived in oss/components/PageTitle/utils and /m had its own `parts` array doing
roughly the same thing. They now share one implementation, with its test.

The surrogate-pair guard is the reason this is worth sharing rather than
reimplementing: a server-persisted session title can arrive cut mid-pair, and the
trailing high surrogate renders as a replacement character. Each host's PageTitle
is now just the <title> element for its router.

Lane: @agenta/shared, below everything else; the oss and mobile rewires ride with
it since the old util is deleted in the same change.
