+++ web/mobile/src/components/PageTitle.tsx
@@ -1,14 +1,21 @@
+import {formatPageTitle} from "@agenta/shared/utils"
 import Head from "next/head"
 
 /**
- * The tab name, in one place.
- *
- * Most specific part first: a phone's tab strip and the app switcher both show only the leading
- * characters, so "approval-flow-test" has to win over the product name. Empty parts drop out, so
- * a screen can pass a name that has not resolved yet without rendering "undefined".
+ * The tab name, in one place — the SAME formatter the desktop uses, so a tab reads identically
+ * on both surfaces: `Section | Entity`, falling back to `Section | Agenta` when a screen names
+ * no entity, and to `Agenta` alone when it names neither.
  */
-export const PageTitle = ({parts}: {parts: (string | null | undefined)[]}) => (
+export const PageTitle = ({
+    title,
+    context,
+}: {
+    /** What this screen is. Omit only on a route that has not resolved yet. */
+    title?: string | null
+    /** What it is about — the agent, the template. Defaults to the product name. */
+    context?: string | null
+}) => (
     <Head>
-        <title>{[...parts.filter(Boolean), "Agenta"].join(" · ")}</title>
+        <title>{formatPageTitle(title, context)}</title>
     </Head>
 )
